third_party/WebKit/LayoutTests/external/wpt/html/semantics/forms/the-input-element/image-click-form-data-expected.txt

This is a testharness.js-based test.
FAIL Image submit button should not add extra form data if 'value' attribute is present with non-empty value assert_equals: expected "http://web-platform.test:8001/html/semantics/forms/the-input-element/resources/image-submit-click.html?name.x=0&name.y=0" but got "http://web-platform.test:8001/html/semantics/forms/the-input-element/resources/image-submit-click.html?name.x=0&name.y=0&name=value"
Harness: the test ran to completion.
